requirements.txt
fastapi
uvicorn
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
psycopg2-binary
sqlalchemy
